+++ precargas.h
@@ -2,12 +2,12 @@
 #define PRECARGAS_H_INCLUDED
 #include "constants.h"
 
-void cargarOscilador(int** matriz, int i, int j);
+void cargarOscilador(int8_t** matriz, int i, int j);
 
-void cargarOscilador2(int** matriz, int i, int j);
+void cargarOscilador2(int8_t** matriz, int i, int j);
 
-void cargarGenShip(int** matriz, int i, int j);
+void cargarGenShip(int8_t** matriz, int i, int j);
 
-void cargarShip(int** matriz, int i, int j);
+void cargarShip(int8_t** matriz, int i, int j);
 
 #endif // PRECARGAS_H_INCLUDED
